Add cfiles_error function for easier error formating

+++ src/utils.hpp
@@ -14,8 +14,8 @@ namespace chemfiles {
 /// Convert a string to double
 double string2double(const std::string& string);
 
-/// Convert a string to long
-long string2long(const std::string& string);
+/// Convert a string to an integer
+long long string2long(const std::string& string);
 
 /// Get a string describing the full version of cfiles
 std::string full_version();
